requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
yfinance
requests
python-dotenv
